+++ src/main/java/com/binance/connector/client/impl/SpotClientImpl.java
@@ -23,42 +23,40 @@ import com.binance.connector.client.impl.spot.SubAccount;
 import com.binance.connector.client.impl.spot.Trade;
 import com.binance.connector.client.impl.spot.UserData;
 import com.binance.connector.client.impl.spot.Wallet;
+import com.binance.connector.client.utils.HmacSignatureGenerator;
+import com.binance.connector.client.utils.SignatureGenerator;
 
 
 public class SpotClientImpl implements SpotClient {
     private final String apiKey;
-    private final String secretKey;
+    private final SignatureGenerator signatureGenerator;
     private final String baseUrl;
     private boolean showLimitUsage = false;
 
     public SpotClientImpl() {
-        this.apiKey = null;
-        this.secretKey = null;
-        this.baseUrl = DefaultUrls.PROD_URL;
+        this(DefaultUrls.PROD_URL);
     }
 
     public SpotClientImpl(String baseUrl) {
-        this.apiKey = null;
-        this.secretKey = null;
-        this.baseUrl = baseUrl;
+        this("", "", baseUrl);
     }
 
     public SpotClientImpl(String baseUrl, boolean showLimitUsage) {
-        this.apiKey = null;
-        this.secretKey = null;
-        this.baseUrl = baseUrl;
+        this(baseUrl);
         this.showLimitUsage = showLimitUsage;
     }
 
     public SpotClientImpl(String apiKey, String secretKey) {
-        this.apiKey = apiKey;
-        this.secretKey = secretKey;
-        this.baseUrl = DefaultUrls.PROD_URL;
+        this(apiKey, secretKey, DefaultUrls.PROD_URL);
     }
 
     public SpotClientImpl(String apiKey, String secretKey, String baseUrl) {
+        this(apiKey, new HmacSignatureGenerator(secretKey), baseUrl);
+    }
+
+    public SpotClientImpl(String apiKey, SignatureGenerator signatureGenerator, String baseUrl) {
         this.apiKey = apiKey;
-        this.secretKey = secretKey;
+        this.signatureGenerator = signatureGenerator;
         this.baseUrl = baseUrl;
     }
 
@@ -68,45 +66,45 @@ public class SpotClientImpl implements SpotClient {
 
     @Override
     public Blvt createBlvt() {
-        return new Blvt(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Blvt(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public BSwap createBswap() {
-        return new BSwap(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new BSwap(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public C2C createC2C() {
-        return new C2C(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new C2C(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Convert createConvert() {
-        return new Convert(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Convert(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public CryptoLoans createCryptoLoans() {
-        return new CryptoLoans(baseUrl, apiKey, secretKey, showLimitUsage); }
+        return new CryptoLoans(baseUrl, apiKey, signatureGenerator, showLimitUsage); }
 
     @Override
     public Fiat createFiat() {
-        return new Fiat(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Fiat(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Futures createFutures() {
-        return new Futures(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Futures(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public GiftCard createGiftCard() {
-        return new GiftCard(baseUrl, apiKey, secretKey, showLimitUsage); }
+        return new GiftCard(baseUrl, apiKey, signatureGenerator, showLimitUsage); }
 
     @Override
     public Margin createMargin() {
-        return new Margin(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Margin(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
@@ -116,47 +114,47 @@ public class SpotClientImpl implements SpotClient {
 
     @Override
     public Mining createMining() {
-        return new Mining(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Mining(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public NFT createNFT() {
-        return new NFT(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new NFT(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Pay createPay() {
-        return new Pay(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Pay(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public PortfolioMargin createPortfolioMargin() {
-        return new PortfolioMargin(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new PortfolioMargin(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Rebate createRebate() {
-        return new Rebate(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Rebate(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Savings createSavings() {
-        return new Savings(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Savings(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Staking createStaking() {
-        return new Staking(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Staking(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public SubAccount createSubAccount() {
-        return new SubAccount(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new SubAccount(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
     public Trade createTrade() {
-        return new Trade(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Trade(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 
     @Override
@@ -166,6 +164,6 @@ public class SpotClientImpl implements SpotClient {
 
     @Override
     public Wallet createWallet() {
-        return new Wallet(baseUrl, apiKey, secretKey, showLimitUsage);
+        return new Wallet(baseUrl, apiKey, signatureGenerator, showLimitUsage);
     }
 }