+++ src/test/java/unit/market/TestAggTrades.java
@@ -39,7 +39,7 @@ public class TestAggTrades {
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
 
-        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        SpotClientImpl client = new SpotClientImpl(baseUrl);
         assertThrows(BinanceConnectorException.class, () -> client.createMarket().aggTrades(parameters));
     }
 
@@ -52,7 +52,7 @@ public class TestAggTrades {
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
 
-        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        SpotClientImpl client = new SpotClientImpl(baseUrl);
         String result = client.createMarket().aggTrades(parameters);
         assertEquals(MockData.MOCK_RESPONSE, result);
     }
@@ -67,7 +67,7 @@ public class TestAggTrades {
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
 
-        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        SpotClientImpl client = new SpotClientImpl(baseUrl);
         String result = client.createMarket().aggTrades(parameters);
         assertEquals(MockData.MOCK_RESPONSE, result);
     }
@@ -82,7 +82,7 @@ public class TestAggTrades {
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
 
-        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        SpotClientImpl client = new SpotClientImpl(baseUrl);
         String result = client.createMarket().aggTrades(parameters);
         assertEquals(MockData.MOCK_RESPONSE, result);
     }
@@ -99,7 +99,7 @@ public class TestAggTrades {
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
 
-        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        SpotClientImpl client = new SpotClientImpl(baseUrl);
         String result = client.createMarket().aggTrades(parameters);
         assertEquals(MockData.MOCK_RESPONSE, result);
     }